issue#02 cambio edad a byte

+++ src/main/java/com/example/demo/model/dto/PersonaDTO.java
@@ -1,4 +1,4 @@
 package com.example.demo.model.dto;
 
-public record PersonaDTO(String nombre, String apellido, int edad,int color) {
+public record PersonaDTO(String nombre, String apellido, byte edad,int color) {
 }
